# shared/api/openapi.yaml
openapi: 3.1.0
info:
  title: Tộc Phả API
  version: v2
  description: |
    Hợp đồng API của Tộc Phả (bản chốt v2: một gia phả chung, hai vai trò Admin và User).
    File này viết tay và là **nguồn sự thật** (DECISIONS #70): frontend sinh kiểu TypeScript từ đây
    (`npm run gen:api`), backend phải khớp (test hợp đồng từ Đợt 26). Muốn đổi API thì sửa file này trước.

    Khởi tạo từ `/v3/api-docs` của backend Đợt 8, đã bỏ `/api/family/**`. Từ Đợt 39 frontend gọi backend thật cho
    mọi endpoint (đã gỡ lớp giả lập).

    Mọi API nghiệp vụ chỉ nhận tài khoản đã được duyệt (403 `ACCOUNT_NOT_APPROVED`), **trừ** các GET xem công khai
    (thành viên, người thân, cây, sự kiện, lịch, tổng quan; DECISIONS #88): khách và tài khoản chưa duyệt cũng gọi được,
    được đánh dấu `security: [{}, {bearerAuth: []}]`. Lỗi luôn là `ProblemDetail`.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /
    description: 'Cùng origin với frontend (dev: proxy /api sang :8080)'
security:
  - bearerAuth: []
...
paths:
...
  /api/tree:
    get:
      tags:
        - tree
      summary: Toàn bộ cây gia phả
      description: |
        Không cần đăng nhập (khách và tài khoản chưa duyệt cũng xem được, DECISIONS #88). Trả toàn bộ đồ thị (nút và cạnh); **đời không nằm trong response**,
        frontend tự tính theo độ sâu (DECISIONS #34, #60). Cây trống thì cả hai mảng rỗng.

        Quy ước của đồ thị:
        - Ô **thuộc dòng** là ô không xuất hiện trong `spouses[].spouseNodeId`. Ô không có `parentNodeId` là **gốc**
          (có thể có nhiều gốc, mỗi gốc là Đời 01 của cây rời của nó).
        - Ô **vợ/chồng** là ô có trong `spouses[].spouseNodeId`. Ô này không có `parentNodeId` và không có vợ/chồng riêng.
        - Con của một cặp: `parentNodeId` luôn là ô **thuộc dòng**, `coParentNodeId` là ô vợ/chồng của ô đó
          (hoặc `null` khi con chỉ thuộc một mình cha/mẹ). Bấm "+ Con" trên ô vợ/chồng thì `parentNodeId` là ô thuộc dòng của ô đó.
        - `memberId = null` là ô trống (đã gỡ người ra khỏi cây); ô trống vẫn có thể có con và vợ/chồng.
        - Anh em (cùng `parentNodeId`, hoặc cùng là gốc) xếp theo `sortOrder` tăng dần, rồi theo `id`.
      operationId: getTree
      security:
        - {}
        - bearerAuth: []
      responses:
        '200':
          description: Thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TreeResponse'
  /api/tree/roots:
    post:
      tags:
        - tree
      summary: Thêm người gốc
      description: |
        Chỉ Admin. Đưa một thành viên chưa có trên cây vào làm gốc mới (Đời 01). Có thể có nhiều gốc; gốc mới xếp cuối.
        Lỗi: 409 `MEMBER_ALREADY_ON_TREE`.
      operationId: addTreeRoot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TreeMemberInput'
      responses:
        '201':
          description: Đã thêm, trả ô mới
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TreeNode'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/tree/nodes/{id}/children:
    post:
      tags:
        - tree
      summary: Thêm con vào một ô
      description: |
        Chỉ Admin. Con mới xếp cuối các anh em. Cặp cha–mẹ được xác định như sau:
        - bấm trên **ô vợ/chồng** thì cặp là ô đó cùng người thuộc dòng của nó (`coParentNodeId`, nếu gửi, phải là chính ô đó);
        - bấm trên ô thuộc dòng có **≥ 2 vợ/chồng** thì bắt buộc gửi `coParentNodeId`, thiếu thì 409 `TREE_NEEDS_CO_PARENT`;
        - có đúng 1 vợ/chồng thì tự nhận cặp đó; không có thì con chỉ thuộc một mình cha/mẹ.

        Lỗi: 404 `TREE_NODE_NOT_FOUND`; 409 `MEMBER_ALREADY_ON_TREE`, `TREE_INVALID_CO_PARENT`, `TREE_NEEDS_CO_PARENT`.
      operationId: addTreeChild
      parameters:
        - $ref: '#/components/parameters/TreeNodeId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TreeChildInput'
      responses:
        '201':
          description: Đã thêm, trả ô mới
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TreeNode'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/tree/nodes/{id}/spouses:
    post:
      tags:
        - tree
      summary: Thêm vợ/chồng cho một ô
      description: |
        Chỉ Admin. Chỉ có trên ô **thuộc dòng** (ô vợ/chồng không có vợ/chồng riêng). Người mới xếp sau vợ/chồng cuối.
        Lỗi: 404 `TREE_NODE_NOT_FOUND`; 409 `TREE_SPOUSE_NOT_ALLOWED`, `MEMBER_ALREADY_ON_TREE`.
      operationId: addTreeSpouse
      parameters:
        - $ref: '#/components/parameters/TreeNodeId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TreeMemberInput'
      responses:
        '201':
          description: Đã thêm, trả ô vợ/chồng mới
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TreeNode'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/tree/nodes/{id}/parent:
    post:
      tags:
        - tree
      summary: Thêm cha/mẹ cho người gốc
      description: |
        Chỉ Admin. Chỉ có trên ô **thuộc dòng ở Đời 01** (gốc). Người mới trở thành gốc, thế vào chỗ của ô này
        trong danh sách gốc; ô này thành con của người mới nên cả cây rời đó dịch xuống một đời. Không tự đặt giới tính.
        Lỗi: 404 `TREE_NODE_NOT_FOUND`; 409 `TREE_PARENT_ONLY_AT_TOP`, `MEMBER_ALREADY_ON_TREE`.
      operationId: addTreeParent
      parameters:
        - $ref: '#/components/parameters/TreeNodeId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TreeMemberInput'
      responses:
        '201':
          description: Đã thêm, trả ô cha/mẹ mới
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TreeNode'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/tree/nodes/{id}/member:
    put:
      tags:
        - tree
      summary: Điền người vào ô trống
      description: |
        Chỉ Admin. Lỗi: 404 `TREE_NODE_NOT_FOUND`; 409 `TREE_SLOT_NOT_EMPTY` (ô đang có người), `MEMBER_ALREADY_ON_TREE`.
      operationId: fillTreeSlot
      parameters:
        - $ref: '#/components/parameters/TreeNodeId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TreeMemberInput'
      responses:
        '200':
          description: Thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TreeNode'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      tags:
        - tree
      summary: Gỡ người khỏi cây (ô thành ô trống)
      description: |
        Chỉ Admin. Ô thành **ô trống** nằm đúng chỗ cũ, con cháu và vợ/chồng không bị ảnh hưởng; thành viên vẫn còn nguyên.
        Lỗi: 404 `TREE_NODE_NOT_FOUND`; 409 `TREE_SLOT_EMPTY` (ô đã trống).
      operationId: removeTreeMember
      parameters:
        - $ref: '#/components/parameters/TreeNodeId'
      responses:
        '200':
          description: Thành công, trả ô đã trống
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TreeNode'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/tree/nodes/{id}:
    delete:
      tags:
        - tree
      summary: Xóa ô trống
      description: |
        Chỉ Admin. Chỉ xóa được **ô trống** (đã gỡ người), kể cả khi ô còn con hoặc vợ/chồng; nhánh được giữ lại:
        - ô vợ/chồng: bị bỏ, con của cặp đó thành con của một mình người thuộc dòng, thứ tự các vợ/chồng còn lại đánh lại từ 1;
        - ô thuộc dòng có vợ/chồng: vợ/chồng thứ nhất thế vào đúng chỗ (cha/mẹ, cặp cha–mẹ, thứ tự anh em), nhận con cháu và các vợ/chồng còn lại;
        - ô thuộc dòng không có vợ/chồng: con cháu lên thế chỗ, làm con của cha/mẹ của ô bị xóa (hoặc thành gốc), cả nhánh dịch lên một đời.
        Thứ tự anh em được đánh lại từ 1. Lỗi: 404 `TREE_NODE_NOT_FOUND`; 409 `TREE_SLOT_NOT_EMPTY` (ô đang có người).
      operationId: deleteTreeNode
      parameters:
        - $ref: '#/components/parameters/TreeNodeId'
      responses:
        '204':
          description: Đã xóa
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/tree/nodes/{id}/move:
    post:
      tags:
        - tree
      summary: Di chuyển nhánh
      description: |
        Chỉ Admin. Một ô **thuộc dòng** đi kèm vợ/chồng và toàn bộ con cháu, tới làm con của `newParentNodeId`
        (ô thuộc dòng hoặc ô vợ/chồng; cặp cha–mẹ xác định như `POST /api/tree/nodes/{id}/children`) hoặc thành gốc mới
        khi `newParentNodeId` là `null`. Ô được chuyển xếp cuối các anh em ở nơi đến.
        Lỗi: 404 `TREE_NODE_NOT_FOUND`; 409 `TREE_MOVE_LINEAGE_ONLY` (ô vợ/chồng), `TREE_CYCLE` (nơi đến nằm trong chính nhánh),
        `TREE_NEEDS_CO_PARENT`, `TREE_INVALID_CO_PARENT`.
      operationId: moveTreeNode
      parameters:
        - $ref: '#/components/parameters/TreeNodeId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TreeMoveInput'
      responses:
        '200':
          description: Thành công, trả ô đã chuyển
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TreeNode'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/tree/nodes/{id}/order:
    put:
      tags:
        - tree
      summary: Đổi thứ tự anh em
      description: |
        Chỉ Admin. Đổi chỗ ô thuộc dòng với người anh em kề bên (cùng cha/mẹ, hoặc cùng là gốc).
        Lỗi: 404 `TREE_NODE_NOT_FOUND`; 409 `TREE_MOVE_LINEAGE_ONLY` (ô vợ/chồng), `TREE_ORDER_EDGE` (đã ở đầu hoặc cuối hàng).
      operationId: reorderTreeNode
      parameters:
        - $ref: '#/components/parameters/TreeNodeId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TreeOrderInput'
      responses:
        '200':
          description: Thành công, trả ô đã đổi chỗ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TreeNode'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/tree/nodes/{id}/co-parent:
    put:
      tags:
        - tree
      summary: Đổi cặp cha–mẹ của một người con
      description: |
        Chỉ Admin. `coParentNodeId` phải là vợ/chồng của cha/mẹ hiện tại (`parentNodeId`). Cha/mẹ có đúng 1 vợ/chồng thì `null` tự nhận người đó;
        có ≥ 2 vợ/chồng thì bắt buộc chọn (409 `TREE_NEEDS_CO_PARENT`).
        Lỗi: 404 `TREE_NODE_NOT_FOUND`; 409 `TREE_NOT_A_CHILD` (ô không có cha/mẹ), `TREE_INVALID_CO_PARENT`, `TREE_NEEDS_CO_PARENT`.
      operationId: setTreeCoParent
      parameters:
        - $ref: '#/components/parameters/TreeNodeId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TreeCoParentInput'
      responses:
        '200':
          description: Thành công, trả ô đã đổi cặp
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TreeNode'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    TreeNodeId:
      name: id
      in: path
      required: true
      description: '`id` của ô trên cây (không phải id thành viên)'
      schema:
        type: integer
        format: int64
        minimum: 1
...
  responses:
    ValidationError:
      description: 'Dữ liệu không hợp lệ (`VALIDATION_ERROR`, `MALFORMED_REQUEST`, lỗi lịch âm...)'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    Unauthorized:
      description: Chưa đăng nhập, phiên hết hạn hoặc sai thông tin đăng nhập (`UNAUTHENTICATED`, `INVALID_CREDENTIALS`...)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    Forbidden:
      description: Không có quyền (`FORBIDDEN`), tài khoản chưa được duyệt (`ACCOUNT_NOT_APPROVED`) hoặc bị khóa (`ACCOUNT_LOCKED`)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    NotFound:
      description: Không tìm thấy (`ACCOUNT_NOT_FOUND`, `MEMBER_NOT_FOUND`, `EVENT_NOT_FOUND`, `TREE_NODE_NOT_FOUND`, `RELATIVE_NOT_FOUND`, `LINK_REQUEST_NOT_FOUND`)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    Conflict:
      description: 'Xung đột trạng thái (`SELF_ACTION_FORBIDDEN`, `LAST_ADMIN`, `INVALID_ACCOUNT_STATE`, `MEMBER_ON_TREE`, `RELATIVE_EXISTS`, `MEMBER_ALREADY_LINKED`, `ACCOUNT_ALREADY_LINKED`, `LINK_REQUEST_EXISTS`, `LINK_REQUEST_NOT_PENDING`, `NOT_LINKED`, các mã cây `MEMBER_ALREADY_ON_TREE`, `TREE_*`)'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
...
  schemas:
    ProblemDetail:
      type: object
      description: Lỗi theo RFC 9457; `detail` là thông báo tiếng Việt hiển thị được cho người dùng.
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
        code:
          type: string
          description: Mã lỗi máy đọc được, ví dụ `ACCOUNT_NOT_APPROVED`, `MEMBER_NOT_FOUND`.
        errors:
          type: array
          description: Lỗi theo từng trường (chỉ có khi `code = VALIDATION_ERROR` hoặc lỗi lịch âm).
          items:
            $ref: '#/components/schemas/FieldError'
    FieldError:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
      required:
        - field
        - message
...
    TreeMember:
      type: object
      description: Tóm tắt thành viên hiển thị trong ô của cây. Họ tên ghi nguyên văn.
      properties:
        fullName:
          type: string
        gender:
          type:
            - string
            - 'null'
          enum:
            - M
            - F
            - null
          description: Được để trống; ô của người chưa rõ giới tính vẽ trung tính
        avatarUrl:
          type:
            - string
            - 'null'
        labels:
          type: array
          description: Nhãn đặc biệt, ví dụ "Liệt sỹ"
          items:
            type: string
        birthYear:
          type:
            - integer
            - 'null'
          format: int32
        isDeceased:
          type: boolean
        deathYear:
          type:
            - integer
            - 'null'
          format: int32
          description: Năm mất theo dương lịch (nếu biết)
      required:
        - fullName
        - gender
        - avatarUrl
        - labels
        - birthYear
        - isDeceased
        - deathYear
    TreeNode:
      type: object
      description: Một ô trên cây. Đời không có trong DTO (tính theo độ sâu ở frontend).
      properties:
        id:
          type: integer
          format: int64
        memberId:
          type:
            - integer
            - 'null'
          format: int64
          description: '`null` là ô trống. Mỗi thành viên có tối đa một ô trên cây.'
        member:
          description: Tóm tắt thành viên; `null` khi là ô trống (khớp với `memberId`)
          oneOf:
            - $ref: '#/components/schemas/TreeMember'
            - type: 'null'
        parentNodeId:
          type:
            - integer
            - 'null'
          format: int64
          description: Ô cha/mẹ **thuộc dòng**; `null` với gốc và với ô vợ/chồng
        coParentNodeId:
          type:
            - integer
            - 'null'
          format: int64
          description: Ô vợ/chồng của `parentNodeId` cùng sinh ra ô này; `null` khi chỉ thuộc một mình cha/mẹ
        sortOrder:
          type: integer
          format: int32
          description: Thứ tự anh em, nhỏ trước
      required:
        - id
        - memberId
        - member
        - parentNodeId
        - coParentNodeId
        - sortOrder
    TreeSpouse:
      type: object
      description: Quan hệ vợ/chồng của một ô thuộc dòng. Mỗi ô vợ/chồng thuộc đúng một ô thuộc dòng.
      properties:
        nodeId:
          type: integer
          format: int64
          description: Ô thuộc dòng
        spouseNodeId:
          type: integer
          format: int64
          description: Ô vợ/chồng
        order:
          type: integer
          format: int32
          minimum: 1
          description: Thứ tự vợ/chồng (1 = Cả, 2 = Hai...)
      required:
        - nodeId
        - spouseNodeId
        - order
    TreeResponse:
      type: object
      properties:
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/TreeNode'
        spouses:
          type: array
          items:
            $ref: '#/components/schemas/TreeSpouse'
      required:
        - nodes
        - spouses
    TreeMemberInput:
      type: object
      properties:
        memberId:
          type: integer
          format: int64
          minimum: 1
          description: Thành viên chưa có trên cây (mỗi thành viên chỉ có một ô)
      required:
        - memberId
    TreeChildInput:
      type: object
      properties:
        memberId:
          type: integer
          format: int64
          minimum: 1
        coParentNodeId:
          type:
            - integer
            - 'null'
          format: int64
          description: Ô vợ/chồng là mẹ/cha còn lại của con; bắt buộc khi người này có từ 2 vợ/chồng trở lên
      required:
        - memberId
    TreeMoveInput:
      type: object
      properties:
        newParentNodeId:
          type:
            - integer
            - 'null'
          format: int64
          description: Ô đến làm cha/mẹ; `null` là thành gốc mới
        coParentNodeId:
          type:
            - integer
            - 'null'
          format: int64
          description: Ô vợ/chồng là mẹ/cha còn lại; bắt buộc khi ô đến có từ 2 vợ/chồng trở lên
      required:
        - newParentNodeId
    TreeOrderInput:
      type: object
      properties:
        direction:
          type: string
          enum:
            - LEFT
            - RIGHT
          description: Đổi chỗ với người anh em bên trái hoặc bên phải
      required:
        - direction
    TreeCoParentInput:
      type: object
      properties:
        coParentNodeId:
          type:
            - integer
            - 'null'
          format: int64
      required:
        - coParentNodeId
...